Restructure the whole project, gofmt and improve the makefile
