+++ cmd/minikube/cmd/root.go
@@ -22,7 +22,7 @@ import (
 	"os"
 
 	"github.com/spf13/cobra"
-	"k8s.io/minikube/cli/constants"
+	"k8s.io/minikube/pkg/minikube/constants"
 )
 
 var dirs = [...]string{
