+++ cmd/minikube/cmd/root_test.go
@@ -23,7 +23,7 @@ import (
 	"testing"
 
 	"github.com/spf13/cobra"
-	"k8s.io/minikube/cli/constants"
+	"k8s.io/minikube/pkg/minikube/constants"
 )
 
 func makeTempDir() string {
