+++ cmd/minikube/cmd/status.go
@@ -18,33 +18,32 @@ package cmd
 
 import (
 	"fmt"
+	"log"
 	"os"
 
 	"github.com/docker/machine/libmachine"
 	"github.com/spf13/cobra"
-	"k8s.io/minikube/cli/cluster"
-	"k8s.io/minikube/cli/constants"
+	"k8s.io/minikube/pkg/minikube/cluster"
+	"k8s.io/minikube/pkg/minikube/constants"
 )
 
-// deleteCmd represents the delete command
-var deleteCmd = &cobra.Command{
-	Use:   "delete",
-	Short: "Deletes a local kubernetes cluster.",
-	Long: `Deletes a local kubernetes cluster. This command deletes the VM, and removes all
-associated files.`,
+// statusCmd represents the status command
+var statusCmd = &cobra.Command{
+	Use:   "status",
+	Short: "Gets the status of a local kubernetes cluster.",
+	Long:  `Gets the status of a local kubernetes cluster.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("Deleting local Kubernetes cluster...")
 		api := libmachine.NewClient(constants.Minipath, constants.MakeMiniPath("certs"))
 		defer api.Close()
-
-		if err := cluster.DeleteHost(api); err != nil {
-			fmt.Println("Errors occurred deleting machine: ", err)
+		s, err := cluster.GetHostStatus(api)
+		if err != nil {
+			log.Println("Error getting machine status:", err)
 			os.Exit(1)
 		}
-		fmt.Println("Machine deleted.")
+		fmt.Fprintln(os.Stdout, s)
 	},
 }
 
 func init() {
-	RootCmd.AddCommand(deleteCmd)
+	RootCmd.AddCommand(statusCmd)
 }
